pull messages endpoint into its own controller for easier maintainability and remove testing routes not being used
